requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
alembic==1.13.1
